refactor: move useIsRunningOnClient
